requirements.txt
numpy
openpyxl
pandas
plotly
dash
dash-bootstrap-components
dash-ag-grid
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread
gspread-dataframe